requirements.txt
plexapi
Pillow
typer